EditorUpdate
Added Editor mode
Now able in the json edit tab to change the values, this currently allows any changes so not handled the best but more for advanced use cases.

Changed how some of the viewer and buttons work
Load to and from clipboard should now be dealt with
Work will be required on load 2d and 3d jsons as these are only test mimic data currently.